Tracking mass hypotheses command line args are now GEANT PIDs (ints) instead of masses. This makes them easier to work with in the analysis library. In analysis library: Now look at command line input to see which PIDs were used for reconstruction. This is necessary to make sure that hypotheses are not recreated for ones that failed pre-select cuts.

+++ src/libraries/TRACKING/DTrackTimeBased_factory.h
@@ -57,8 +57,8 @@ class DTrackTimeBased_factory:public jana::JFactory<DTrackTimeBased>{
   DTrackFitter *fitter;
   const DParticleID* pid_algorithm;
   vector<DReferenceTrajectory*> rtv;	
-  vector<double> mass_hypotheses_positive;
-  vector<double> mass_hypotheses_negative;
+  vector<int> mass_hypotheses_positive;
+  vector<int> mass_hypotheses_negative;
   size_t MAX_DReferenceTrajectoryPoolSize;
